+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.5.2]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [1.1.0]
+
+### Added
+- extension may be omitted when loading dataset in wildcard mode if dataset doesn't have extra parts
+
+## [1.0.0]
+
+### Modified
+- removed deprecated `_from_` converters
+- removed walrus from test
+- raise warning when fails to set dataset exists to False
+
+## [0.5.3]
+
+### Modified
+
+- HOTFIX: Error no longer raised when logging dimension mismatch in alf.io.load_object
+
+## [0.5.2]
 
 ### Modified